feat: 시간 포맷팅 함수 생성
- util에 추가
- issue화면에 쓰이고 있던 util 함수 import/export 방식 수정
